requirements.txt
Django==1.9
django-braces==1.10.0
django-debug-toolbar==1.6
git+https://github.com/jhuapl-boss/django-oidc.git
Unipath==1.1
